+++ Engine/physics/Integrator.h
@@ -26,7 +26,7 @@ public:
     void setScheme(Scheme scheme);
     Scheme getScheme() const { return integrator_type; }
 
-    void step(AtomStorage& atomStorage, SimBox& box, ForceField& forceField, double dt) const;
+    void step(AtomStorage& atomStorage, SimBox& box, ForceField& forceField, float dt) const;
 
 private:
     using SchemeVariant = std::variant<VerletScheme, KDKScheme, RK4Scheme, LangevinScheme>;
